Moving TraversalNode back to original location
