+++ src/fides/api/graph/node_filters.py
@@ -1,13 +1,15 @@
-from typing import Optional, Protocol
+from typing import TYPE_CHECKING, Optional, Protocol
 
-from fides.api.graph.traversal_node import TraversalNode
 from fides.api.models.policy import Policy
+
+if TYPE_CHECKING:
+    from fides.api.graph.traversal import TraversalNode
 
 
 class NodeFilter(Protocol):
     """Protocol for node filters that determine if nodes should be excluded from traversal"""
 
-    def exclude_node(self, node: TraversalNode) -> bool:
+    def exclude_node(self, node: "TraversalNode") -> bool:
         """Returns True if the node should be excluded from traversal requirements"""
 
 
@@ -20,7 +22,7 @@ class CustomRequestFieldFilter:
     as part of the main graph. This is a targeted workaround for now.
     """
 
-    def exclude_node(self, node: TraversalNode) -> bool:
+    def exclude_node(self, node: "TraversalNode") -> bool:
         return bool(node.node.collection.custom_request_fields())
 
 
@@ -33,7 +35,7 @@ class PolicyDataCategoryFilter:
     def __init__(self, policy: Optional[Policy]):
         self.policy = policy
 
-    def exclude_node(self, node: TraversalNode) -> bool:
+    def exclude_node(self, node: "TraversalNode") -> bool:
         if not self.policy:
             return False
         return not self.policy.applies_to(node)
