+++ src/fides/api/task/deprecated_graph_task.py
@@ -17,9 +17,9 @@ from fides.api.graph.config import (
 from fides.api.graph.graph import DatasetGraph
 from fides.api.graph.traversal import (
     Traversal,
+    TraversalNode,
     log_traversal_error_and_update_privacy_request,
 )
-from fides.api.graph.traversal_node import TraversalNode
 from fides.api.models.connectionconfig import ConnectionConfig
 from fides.api.models.policy import Policy
 from fides.api.models.privacy_request import PrivacyRequest
